requirements.txt
Django>=6.0
gpxpy>=1.5.0
pillow>=10.0.0
folium>=0.14.0
django-storages>=1.14.0
boto3>=1.28.0
requests>=2.31.0
matplotlib>=3.7.0
playwright>=1.40.0
